Move hashes inside pluginsById
Otherwise it would be difficult to keep the whole state of the plugin
together, which would make saving more complex
